Moved Feature to Dataset model file.
